# perf(filter): add cache for pattern matching handler