--- a/filter/patterns_storage.go
+++ b/filter/patterns_storage.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/moira-alert/moira/clock"
 
-	lruCache "github.com/hashicorp/golang-lru/v2"
+	lrucache "github.com/hashicorp/golang-lru/v2"
 	"github.com/moira-alert/moira"
 	"github.com/moira-alert/moira/metrics"
 )
@@ -32,7 +32,7 @@ type PatternStorage struct {
 	PatternIndex            atomic.Value
 	SeriesByTagPatternIndex atomic.Value
 	compatibility           Compatibility
-	patternMatchingCache    *lruCache.Cache[string, *patternMatchingCacheItem]
+	patternMatchingCache    *lrucache.Cache[string, *patternMatchingCacheItem]
 }
 
 // NewPatternStorage creates new PatternStorage struct.
@@ -43,7 +43,7 @@ func NewPatternStorage(
 	logger moira.Logger,
 	compatibility Compatibility,
 ) (*PatternStorage, error) {
-	patternMatchingCache, err := lruCache.New[string, *patternMatchingCacheItem](cfg.PatternMatchingCacheSize)
+	patternMatchingCache, err := lrucache.New[string, *patternMatchingCacheItem](cfg.PatternMatchingCacheSize)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create new lru pattern matching cache: %w", err)
 	}