--- a/filter/patterns_storage_test.go
+++ b/filter/patterns_storage_test.go
@@ -27,20 +27,26 @@ func TestProcessIncomingMetric(t *testing.T) {
 	database := mock_moira_alert.NewMockDatabase(mockCtrl)
 	logger, _ := logging.ConfigureLog("stdout", "warn", "test", true)
 
+	patternStorageCfg := PatternStorageConfig{
+		TagsRegexCacheSize: 100,
+	}
+
 	Convey("Create new pattern storage, GetPatterns returns error, should error", t, func() {
 		database.EXPECT().GetPatterns().Return(nil, fmt.Errorf("some error here"))
 		filterMetrics := metrics.ConfigureFilterMetrics(metrics.NewDummyRegistry())
-		_, err := NewPatternStorage(database, filterMetrics, logger, Compatibility{AllowRegexLooseStartMatch: true})
-		So(err, ShouldBeError, fmt.Errorf("some error here"))
+		_, err := NewPatternStorage(patternStorageCfg, database, filterMetrics, logger, Compatibility{AllowRegexLooseStartMatch: true})
+		So(err, ShouldBeError, fmt.Errorf("failed to refresh pattern storage: some error here"))
 	})
 
 	database.EXPECT().GetPatterns().Return(testPatterns, nil)
 	patternsStorage, err := NewPatternStorage(
+		patternStorageCfg,
 		database,
 		metrics.ConfigureFilterMetrics(metrics.NewDummyRegistry()),
 		logger,
 		Compatibility{AllowRegexLooseStartMatch: true},
 	)
+
 	systemClock := mock_clock.NewMockClock(mockCtrl)
 	systemClock.EXPECT().Now().Return(time.Date(2009, 2, 13, 23, 31, 30, 0, time.UTC)).AnyTimes()
 	patternsStorage.clock = systemClock