--- a/go.mod
+++ b/go.mod
@@ -48,6 +48,7 @@ require (
 require github.com/prometheus/common v0.37.0
 
 require (
+	github.com/hashicorp/golang-lru/v2 v2.0.7
 	github.com/mattermost/mattermost/server/public v0.1.1
 	github.com/mitchellh/mapstructure v1.5.0
 	github.com/moira-alert/blackfriday-slack v0.1.2