# Add pattern matching cache

Added pattern matching handler caching to make GC life a little easier. Also added Refresh pattern storage benchmarks

On prod data with the right cache size:
```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPatternStorageProd/small_cache-12     167404983     171102503     +2.21%
BenchmarkPatternStorageProd/big_cache-12       167404983     122361363     -26.91%

benchmark                                      old allocs     new allocs     delta
BenchmarkPatternStorageProd/small_cache-12     740377         759268         +2.55%
BenchmarkPatternStorageProd/big_cache-12       740377         353399         -52.27%

benchmark                                      old bytes     new bytes     delta
BenchmarkPatternStorageProd/small_cache-12     59553850      60456515      +1.52%
BenchmarkPatternStorageProd/big_cache-12       59553850      29123259      -51.10%
```

On test data:
```
> go test -v -benchmem -memprofile=refresh.mem.out -cpuprofile=refresh.cpu.out -count=1 -benchtime=10s -bench=^BenchmarkPatternStorage . -run=^$ . | tee refresh_bench.txt

goos: linux
goarch: amd64
pkg: github.com/moira-alert/moira/perfomance_tests/filter
cpu: AMD Ryzen 5 5500U with Radeon Graphics         
BenchmarkPatternStorageRefresh
BenchmarkPatternStorageRefresh/small_cache
BenchmarkPatternStorageRefresh/small_cache-12               1140          10600212 ns/op         3580941 B/op   27984 allocs/op
BenchmarkPatternStorageRefresh/big_cache
BenchmarkPatternStorageRefresh/big_cache-12                 2172           5350007 ns/op          415106 B/op    4913 allocs/op
PASS
ok      github.com/moira-alert/moira/perfomance_tests/filter    25.613s
```

You can see that the cache increases speed and also greatly reduces the number of allocations